+++ src/main/java/blackjack/domain/card/Letter.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Letter {
     ACE("A", 1),
