+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
+
+import blackjack.domain.card.Deck;
 
 import java.util.Collections;
 import java.util.List;
