+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.result;
+
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Participant;
+import blackjack.domain.player.Player;
 
 import java.util.*;
