+++ src/test/java/blackjack/domain/DealerTest.java
@@ -1,5 +1,9 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Letter;
+import blackjack.domain.card.Suit;
+import blackjack.domain.player.Dealer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
