+++ src/test/java/blackjack/domain/NameTest.java
@@ -1,8 +1,8 @@
 package blackjack.domain;
 
+import blackjack.domain.player.Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
