+++ src/test/java/blackjack/domain/ScoreTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.result.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
